Update content.md
The pseudo globals that start w/ 2 underscores were having the 2 underscores interpreted as markdown. Since this page wanted the globals to be marked as code anyway, I've added that markup around the `__filename`, `__dirname` and 'require()` bullet entries. (for some reason preview isn't working for me, so I think I've got this correct, but couldn't verify it.